add upper/lower limit which the get_covering can go. And add a config parser

+++ simp_and_store/load_refactor/dbhelper.py
@@ -1,11 +1,17 @@
 '''Dbhelper module, this module contains the abstraction of storing data into database'''
 
 from uuid import uuid1 as timeuuid
+import ctypes
+import atexit
 import geojson
 import geohelper
-import atexit
 from cassandra.cluster import Cluster
+import cfgparser
 
+
+TRUNCATE_TABLE_WHEN_START = True
+
+# TODO: the following parameters need to be added to config.yml
 # CLUSTER_LIST = ['192.168.1.10']
 CLUSTER_LIST = None
 
@@ -19,9 +25,11 @@ PREPARED_INSERT = SESSION.prepare(PS_INSERT)
 def insert_by_covering(cellid, feature):
     '''Given the covering region, store the given feature into the database'''
     osm_id = feature['id']
+    s2_id = ctypes.c_long(cellid.id()).value
     feature_str = geojson.dumps(feature)
     insert_by_covering.handle = SESSION.execute_async(
-        PREPARED_INSERT, (cellid.level(), cellid.id(), timeuuid(), osm_id, feature_str))
+        PREPARED_INSERT, (cellid.level(), s2_id, timeuuid(), osm_id, feature_str))
+
 
 def insert_by_bboxes(bboxes, feature):
     '''Given the bboxes of the feature and the feature itself, store the given feature into the
@@ -32,19 +40,28 @@ def insert_by_bboxes(bboxes, feature):
             insert_by_covering(covering, feature)
 
 
-def initialize():
+def __initialize():
     '''DO NOT CALL THIS FUNCTION. Initializing the module. '''
-    SESSION.execute('TRUNCATE slave')
+    if TRUNCATE_TABLE_WHEN_START:
+        SESSION.execute('TRUNCATE slave')
 
 
-def before_exit():
+def __before_exit():
     '''Wait to ensure that all insertion has been into the table'''
-    print 'Waiting for database to finish up...'
-    insert_by_covering.handle.result()
+    if insert_by_covering.handle is not None:
+        print 'Waiting for database to finish up...'
+        insert_by_covering.handle.result()
+
+
+def __load_config():
+    cfg = cfgparser.load_module('dbhelper')
+    global TRUNCATE_TABLE_WHEN_START
+    TRUNCATE_TABLE_WHEN_START = cfg['truncate table when start']
+
 
 
 ###########################################
 ##############  INITIALIZE ################
 insert_by_covering.handle = None
-initialize()
-atexit.register(before_exit)
+atexit.register(__before_exit)
+__initialize()
